[jsk_robot_startup] Use common odometry utilities in ParticleOdometry
